backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
python-dotenv==1.0.0